LayoutTests/platform/chromium/svg/text/caret-in-svg-text-expected.txt

Sheriff Woody
שדגש

Failure. Was: 20,15,0,19, expected: 20,566,0,18
Failure. Was: 109,15,0,19, expected: 114,566,0,18
Failure. Was: 20,35,0,19, expected: 58,546,0,18
Failure. Was: 52,35,0,19, expected: 20,546,0,18
